requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
arrow==1.2.3
arxiv==1.4.4
asgiref==3.6.0
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
blobfile==2.0.1
build==0.10.0
CacheControl==0.12.11
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cleo==2.0.1
click==8.1.3
crashtest==0.4.1
cryptography==40.0.1
cssselect==1.2.0
cssutils==2.6.0
dataclasses-json==0.5.7
decorator==5.1.1
distlib==0.3.6
Django==4.2
docx2txt==0.8
dulwich==0.21.3
executing==1.2.0
fastapi==0.95.0
feedparser==6.0.10
filelock==3.11.0
Flask==2.2.3
flask-crontab==0.1.2
Flask-SQLAlchemy==3.0.3
frozenlist==1.3.3
greenlet==2.0.2
grpcio==1.53.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
html5lib==1.1
httpcore==0.16.3
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.3.0
iniconfig==2.0.0
installer==0.7.0
ipython==8.12.0
itsdangerous==2.1.2
jaraco.classes==3.2.3
jedi==0.18.2
jeepney==0.8.0
Jinja2==3.1.2
jsonschema==4.17.3
keyring==23.13.1
lockfile==0.12.2
lxml==4.9.2
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib-inline==0.1.6
mmh3==3.1.0
more-itertools==9.1.0
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.24.2
openai==0.27.4
openapi-schema-pydantic==1.2.4
packaging==23.0
pandas==2.0.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pkginfo==1.9.6
platformdirs==2.6.2
pluggy==1.0.0
poetry==1.4.2
poetry-core==1.5.2
poetry-plugin-export==1.3.0
premailer==3.10.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psycopg2==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pycryptodomex==3.17
pydantic==1.10.7
Pygments==2.15.0
PyJWT==2.6.0
pymilvus==2.2.6
pynliner==0.8.0
pypdf==3.7.1
PyPDF2==3.0.1
pyproject_hooks==1.0.0
pyrsistent==0.19.3
pytest==7.3.0
python-dateutil==2.8.2
python-multipart==0.0.6
python-pptx==0.6.21
pytz==2023.3
PyYAML==6.0
rapidfuzz==2.15.1
regex==2023.3.23
requests==2.28.2
requests-toolbelt==0.10.1
rfc3986==1.5.0
SecretStorage==3.3.3
sgmllib3k==1.0.0
shellingham==1.5.0.post1
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
SQLAlchemy==2.0.9
sqlparse==0.4.3
stack-data==0.6.2
starlette==0.26.1
tenacity==8.2.2
tiktoken==0.3.3
tomlkit==0.11.7
tqdm==4.65.0
traitlets==5.9.0
trove-classifiers==2023.3.9
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
ujson==5.7.0
urllib3==1.26.15
uuid==1.30
uvicorn==0.21.1
virtualenv==20.21.0
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==2.2.3
XlsxWriter==3.0.9
yagmail==0.15.293
yarl==1.8.2
zipp==3.15.0
